requirements.txt
blockcypher==1.0.93
pandas==1.4.2
matplotlib==3.5.1
scipy==1.8.0
python-binance==1.0.16
numpy==1.22.3
pytz